fix: fetch all the domains and the resource_record_count in one separate query
Doing `select_related` and `prefetch_related` for the domain objects resulted in having to calculate the `resource_record_count` for all the nested domain objects inside a node (didn't matter if it was the same domain object across all the nodes or a different one). So, while it's true that we weren't doing any additional queries (because we already fetched all the objects in the `prefetch_related` and `select_related`), we still had to some loops on the python side. Repeat this for N nodes and you get a high waste of CPU.

This commit introduces a new approach of doing it:
- remove the domain and domain-related tables from the `select_related` and `prefetch_related`
- have a single query to get all the domains together with the infamous `resource_record_count`
- assign the domain object to every node by looking at the node's domain_id

Resolves: LP: #2098967