fix(dhcp): create unique host entries in dhcpd.conf for interfaces with same MAC
The solution is leveraging the fact that host `hostname` is just an identifier for dhcpd configuration that can be anything. For interfaces that share the same MAC e.g., VLAN interfaces, which have fixed IP addresses on different VLANs, this change create a host block per interface.

For reserved IPs, to create name uniqueness in the host block hostname, a prefix "rsvd-{reserved_ip.id}-" is added to the MAC address of the reservation.

Resolves LP:2134484